new stuff
I've added some new functionality but most of it isnt working.

In order to try and handle the different components Im trying to create a meta graph in order to do vertex grouping on it.

I'm also now considering the max differnce on the graph nodes. This gives an idea of what the overall coherance of the graph is